# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LottoLab API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/historical-results/runs": {
      "get": {
        "tags": [
          "historical-results"
        ],
        "summary": "List Historical Runs",
        "operationId": "listHistoricalRuns",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalRunPageResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/historical-results/runs/{run_id}/strategies": {
      "get": {
        "tags": [
          "historical-results"
        ],
        "summary": "List Historical Run Strategies",
        "operationId": "listHistoricalRunStrategies",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Run Id"
            }
          },
          {
            "name": "ticket_count",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TicketCount"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalStrategySummaryListResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/historical-results/runs/{run_id}/replay": {
      "get": {
        "tags": [
          "historical-results"
        ],
        "summary": "List Historical Run Replay Portfolios",
        "operationId": "listHistoricalRunReplayPortfolios",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Run Id"
            }
          },
          {
            "name": "strategy_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Strategy Id"
            }
          },
          {
            "name": "ticket_count",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TicketCount"
            }
          },
          {
            "name": "m4plus_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "M4Plus Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalReplayPageResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/historical-results/portfolios/{portfolio_id}": {
      "get": {
        "tags": [
          "historical-results"
        ],
        "summary": "Get Historical Portfolio",
        "operationId": "getHistoricalPortfolio",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Portfolio Id"
            }
          },
          {
            "name": "ticket_count",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TicketCount"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalPortfolioView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ApiErrorResponse"
      },
      "ApiValidationErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DrawImportPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/RequestValidationIssueView"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ApiValidationErrorResponse"
      },
...
      "HistoricalPortfolioView": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "title": "Portfolio Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "strategy_snapshot_id": {
            "type": "string",
            "title": "Strategy Snapshot Id"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "effective_strategy_id": {
            "type": "string",
            "title": "Effective Strategy Id"
          },
          "strategy_version": {
            "type": "string",
            "title": "Strategy Version"
          },
          "replicate": {
            "type": "integer",
            "title": "Replicate"
          },
          "constructor_identifier": {
            "type": "string",
            "title": "Constructor Identifier"
          },
          "source_record_locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Record Locator"
          },
          "portfolio_sha256": {
            "type": "string",
            "title": "Portfolio Sha256"
          },
          "prefix10_sha256": {
            "type": "string",
            "title": "Prefix10 Sha256"
          },
          "prefix15_sha256": {
            "type": "string",
            "title": "Prefix15 Sha256"
          },
          "target_draw": {
            "$ref": "#/components/schemas/HistoricalDrawIdentityView"
          },
          "cutoff_draw": {
            "$ref": "#/components/schemas/HistoricalDrawIdentityView"
          },
          "requested_ticket_count": {
            "type": "integer",
            "title": "Requested Ticket Count"
          },
          "m4plus": {
            "type": "boolean",
            "title": "M4Plus"
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/HistoricalTicketView"
            },
            "type": "array",
            "title": "Tickets"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "run_id",
          "strategy_snapshot_id",
          "strategy_id",
          "effective_strategy_id",
          "strategy_version",
          "replicate",
          "constructor_identifier",
          "source_record_locator",
          "portfolio_sha256",
          "prefix10_sha256",
          "prefix15_sha256",
          "target_draw",
          "cutoff_draw",
          "requested_ticket_count",
          "m4plus",
          "tickets"
        ],
        "title": "HistoricalPortfolioView"
      },
...
      "HistoricalReplayPageResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "ticket_count": {
            "type": "integer",
            "title": "Ticket Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoricalPortfolioView"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "strategy_id",
          "ticket_count",
          "items",
          "total_count",
          "limit",
          "offset"
        ],
        "title": "HistoricalReplayPageResponse"
      },
      "HistoricalRunPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoricalRunView"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "limit",
          "offset"
        ],
        "title": "HistoricalRunPageResponse"
      },
      "HistoricalRunView": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "import_identity_sha256": {
            "type": "string",
            "title": "Import Identity Sha256"
          },
          "manifest_sha256": {
            "type": "string",
            "title": "Manifest Sha256"
          },
          "contract_version": {
            "type": "string",
            "title": "Contract Version"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "source_repository": {
            "type": "string",
            "title": "Source Repository"
          },
          "source_commit_oid": {
            "type": "string",
            "title": "Source Commit Oid"
          },
          "source_artifact_sha256": {
            "type": "string",
            "title": "Source Artifact Sha256"
          },
          "dataset_identity": {
            "type": "string",
            "title": "Dataset Identity"
          },
          "dataset_sha256": {
            "type": "string",
            "title": "Dataset Sha256"
          },
          "legacy_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacy Run Id"
          },
          "lottery_type": {
            "type": "string",
            "title": "Lottery Type"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "completed_at": {
            "type": "string",
            "title": "Completed At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "strategy_count": {
            "type": "integer",
            "title": "Strategy Count"
          },
          "draw_count": {
            "type": "integer",
            "title": "Draw Count"
          },
          "portfolio_count": {
            "type": "integer",
            "title": "Portfolio Count"
          },
          "is_idempotent_replay": {
            "type": "boolean",
            "title": "Is Idempotent Replay"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "import_identity_sha256",
          "manifest_sha256",
          "contract_version",
          "source_kind",
          "source_repository",
          "source_commit_oid",
          "source_artifact_sha256",
          "dataset_identity",
          "dataset_sha256",
          "legacy_run_id",
          "lottery_type",
          "started_at",
          "completed_at",
          "status",
          "strategy_count",
          "draw_count",
          "portfolio_count",
          "is_idempotent_replay"
        ],
        "title": "HistoricalRunView"
      },
      "HistoricalStrategySummaryListResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "ticket_count": {
            "type": "integer",
            "title": "Ticket Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoricalStrategySummaryView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "ticket_count",
          "items"
        ],
        "title": "HistoricalStrategySummaryListResponse"
      },
...
      "RequestValidationIssueView": {
        "properties": {
          "location": {
            "type": "string",
            "title": "Location"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "location",
          "type"
        ],
        "title": "RequestValidationIssueView"
      },
...
      "TicketCount": {
        "type": "integer",
        "enum": [
          10,
          15,
          20
        ],
        "title": "TicketCount",
        "description": "The closed ticket_count set (10/15/20).\n\nA bare ``Literal[10, 15, 20]`` does not coerce a query string like\n\"10\" to the matching int literal and always 422s; ``IntEnum`` does\ncoerce correctly and still round-trips as a plain int in responses."
      },
...
    }
  }
}